Create a minimum viable concept for the frontend app
